Chore: web 패키지를 external로 이동
