+++ order-service/src/main/java/com/logistic/order/adapter/in/external/web/request/CreateOrderRequest.java
@@ -1,4 +1,4 @@
-package com.logistic.order.adapter.in.web.request;
+package com.logistic.order.adapter.in.external.web.request;
 
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotEmpty;
