+++ order-service/src/main/java/com/logistic/order/adapter/in/external/web/response/CreateOrderResponse.java
@@ -1,14 +1,14 @@
-package com.logistic.order.adapter.in.web.response;
+package com.logistic.order.adapter.in.external.web.response;
 
 import java.util.List;
 
-public record UpdateOrderResponse(
+public record CreateOrderResponse(
     Long id,
     Long sellerId,
     Long buyerId,
     String memo,
     String status,
-    List<OrderProduct> orderProductList) {
+    List<OrderProduct> orderProducts) {
 
   public record OrderProduct(
       Long productId,
