+++ order-service/src/main/java/com/logistic/order/adapter/in/external/web/response/SearchOrderResponse.java
@@ -1,14 +1,14 @@
-package com.logistic.order.adapter.in.web.response;
+package com.logistic.order.adapter.in.external.web.response;
 
 import java.util.List;
 
-public record CreateOrderResponse(
+public record SearchOrderResponse(
     Long id,
     Long sellerId,
     Long buyerId,
     String memo,
     String status,
-    List<OrderProduct> orderProducts) {
+    List<OrderProduct> orderProductList) {
 
   public record OrderProduct(
       Long productId,
